requirements.txt
flask>=3.0.0
flask_sqlalchemy>=3.1.1
tensorflow>=2.13.0
numpy>=1.24.0
pillow>=10.0.0
werkzeug>=3.0.0

